Findall Conaluido

+++ src/main/java/br/com/linx/cardValidator/repository/BinTypeRepository.java
@@ -1,13 +1,12 @@
 package br.com.linx.cardValidator.repository;
 
 import br.com.linx.cardValidator.model.BinType;
-import br.com.linx.cardValidator.model.Brand;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
 public interface BinTypeRepository extends JpaRepository<BinType, Integer> {
 
-    List<BinType> findByIdType(int id_type);
+    List<BinType> findByIdType(Long id_type);
 
 }
