+++ src/main/java/br/com/linx/cardValidator/repository/BrandRespository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 public interface BrandRespository extends JpaRepository<Brand, Integer> {
 
-    List<Brand> findByIdBrand(int id_brand);
+    List<Brand> findByIdBrand(Long id_brand);
